video_processing/requirements.txt
opencv-python==4.10.0.84
opencv-contrib-python==4.10.0.84
openai==1.61.0
python-dotenv==1.0.1
numpy==1.26.4